# src/main/resources/swagger/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:10080/backend
    description: Backend
...
paths:
...
  /api/v1/fire-forecast-indexes:
    get:
      tags:
        - FFWIController
      summary: Endpoint для получения всей информации об КППО
      description: Получение списка КППО для отображения на странице
      operationId: getAllFfwi
      responses:
        '200':
          description: Успешный ответ на запрос
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FFWIResponseDto'
        '401':
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: Пример ответа
                  value:
                    message: Пользователь не авторизован
                    code: 401
        '500':
          description: Ошибка на сервере
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                example1:
                  description: Пример ответа на 500
                  value:
                    message: Произошла ошибка на стороне сервера
                    code: 500
    post:
      tags:
        - FFWIController
      summary: Endpoint для добавления нового КППО
      description: Добавление КППО в базу данных
      operationId: createFfwi
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FFWICreateRequestDto'
            examples:
              example1:
                value:
                  name: КППО Свердловой
                  formula: $K_v * ($T_d - $T_o) * $T_d
                  weatherDataParams:
                    - id: 28723f7b-7798-4e34-9569-e46a255aab3d
                      formulaParameter: $K_v
                    - id: 28723f7b-7798-4e34-9569-e46a255aab3d
                      formulaParameter: $T_d
                    - id: 28723f7b-7798-4e34-9569-e46a255aab3d
                      formulaParameter: $T_o

      responses:
        '201':
          description: Успешное создание КППО
        '400':
          description: Ошибка в составлении запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                example1:
                  description: Пример неправильного запроса
                  value:
                    message: Название КППО не соответствует формату
                    code: 400
                example2:
                  description: Формула не соответствует грамматике
                  value:
                    message: Формула не соответствует грамматике
                    code: 400
        '401':
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: Пример ответа
                  value:
                    message: Пользователь не авторизован
                    code: 401
        '403':
          description: Пользователю не хватает прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: Пример ответа
                  value:
                    message: Недостаточно прав для выполнения операции
                    code: 403
        '500':
          description: Ошибка на сервере
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                example1:
                  description: Пример ответа на 500
                  value:
                    message: Произошла ошибка на стороне сервера
                    code: 500
    patch:
      tags:
        - FFWIController
      summary: Обновить информацию о КППО
      description: Обновление информации о КППО и связанных с ним сущностях
      operationId: patchFfwi
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FFWIResponseDto'
      responses:
        '204':
          description: Успешный запрос
        '400':
          description: Ошибка в составлении запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                example1:
                  description: Пример неправильного запроса
                  value:
                    message: Название КППО не соответствует формату
                    code: 400
                example2:
                  description: Формула не соответствует грамматике
                  value:
                    message: Формула не соответствует грамматике
                    code: 400
        '401':
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: Пример ответа
                  value:
                    message: Пользователь не авторизован
                    code: 401
        '403':
          description: Пользователю не хватает прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: Пример ответа
                  value:
                    message: Недостаточно прав для выполнения операции
                    code: 403
        '500':
          description: Ошибка на сервере
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                example1:
                  description: Пример ответа на 500
                  value:
                    message: Произошла ошибка на стороне сервера
                    code: 500
  /api/v1/fire-forecase-indexes/{id}:
    parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
          format: uuid
          description: ID КППО
    get:
      tags:
        - FFWIController
      summary: Получить КППО по id
      description: получение КППО по id
      operationId: getFfwiById
      responses:
        '200':
          description: Успешный ответ
          content:
            appplication/json:
              schema:
                $ref: '#/components/schemas/FFWIResponseDto'
        '400':
          description: Отсутствие id в запросе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: Пример ответа
                  value:
                    message: Отсутствует параметр id в запросе
                    code: 400
        '401':
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: Пример ответа
                  value:
                    message: Пользователь не авторизован
                    code: 401
        '404':
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: КППО не был найден в БД
                  value:
                    message: Запись о КППО 1993dbee-eb94-46cb-aa9d-5b166b1c6659 нет в системе
                    code: 404
        '500':
          description: Ошибка на сервере
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                example1:
                  description: Пример ответа на 500
                  value:
                    message: Произошла ошибка на стороне сервера
                    code: 500
    delete:
      tags:
        - FFWIController
      summary: Удалние КППО по id
      description: Удаление информации об КППО по id
      operationId: deleteFfwiById
      responses:
        '200':
          description: Успешное выполнение запроса
        '400':
          description: Отсутствие id в запросе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: Пример ответа
                  value:
                    message: Отсутствует параметр id в запросе
                    code: 400
                example2:
                  description: Пример ответа 2
                  value:
                    message: КППО связан с другими записями
                    code: 400
        '401':
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: Пример ответа
                  value:
                    message: Пользователь не авторизован
                    code: 401
        '404':
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: КППО не был найден в БД
                  value:
                    message: Запись о КППО 1993dbee-eb94-46cb-aa9d-5b166b1c6659 нет в системе
                    code: 404
        '500':
          description: Ошибка на сервере
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                example1:
                  description: Пример ответа на 500
                  value:
                    message: Произошла ошибка на стороне сервера
                    code: 500
...
  /api/v1/fire-weather-indexes/{id}:
    parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
          format: uuid
          description: ID КПО
    get:
      tags:
        - FWIController
      summary: Получение конкретной записи
      description: Получение конкретной записи об КПО
      operationId: getById
      responses:
        '200':
          description: Успешный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FWIResponseDto'
        '400':
          description: Отсутствие id в запросе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: Пример ответа
                  value:
                    message: Отсутствует параметр id в запросе
                    code: 400
        '401':
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: Пример ответа
                  value:
                    message: Пользователь не авторизован
                    code: 401
        '404':
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: КПО не был найден в БД
                  value:
                    message: Запись о КПО 1993dbee-eb94-46cb-aa9d-5b166b1c6659 нет в системе
                    code: 404
        '500':
          description: Ошибка на сервере
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                example1:
                  description: Пример ответа на 500
                  value:
                    message: Произошла ошибка на стороне сервера
                    code: 500
    delete:
      tags:
        - FWIController
      summary: Удалить запись
      description: Удаление записи по КПО
      operationId: deleteById
      responses:
        '200':
          description: Успешное выполнение запроса
        '400':
          description: Отсутствие id в запросе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: Пример ответа
                  value:
                    message: Отсутствует параметр id в запросе
                    code: 400
                example2:
                  description: Пример ответа 2
                  value:
                    message: КППО связан с другими записями
                    code: 400
        '401':
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: Пример ответа
                  value:
                    message: Пользователь не авторизован
                    code: 401
        '404':
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: КПО не был найден в БД
                  value:
                    message: Запись о КПО 1993dbee-eb94-46cb-aa9d-5b166b1c6659 нет в системе
                    code: 404
        '500':
          description: Ошибка на сервере
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                example1:
                  description: Пример ответа на 500
                  value:
                    message: Произошла ошибка на стороне сервера
                    code: 500
  /api/v1/precipitation-records:
    get:
      tags:
        - PrecipitationRecordController
      summary: Получить записи по таблице осадков
      description: Получение всех записей таблицы осадков
      operationId: getPrecipitationRecords
      responses:
        '200':
          description: Успешное получение всех записей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PrecipitationRecordResponseDto'
        '401':
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: Пример ответа
                  value:
                    message: Пользователь не авторизован
                    code: 401
        '500':
          description: Ошибка на сервере
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                example1:
                  description: Пример ответа на 500
                  value:
                    message: Произошла ошибка на стороне сервера
                    code: 500
    post:
      tags:
        - PrecipitationRecordController
      summary: Добавить запись по таблице осадков
      description: Добавление новой записи по природному осадку в БД
      operationId: addPrecipitationRecord
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrecipitationRecordCreateDto'
      responses:
        '201':
          description: Успешное добавление записи
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrecipitationRecordResponseDto'
        '400':
          description: Ошибка в составлении запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                example1:
                  description: Пример неправильного запроса
                  value:
                    message: Минимальная граница больше максимальной
                    code: 400
                example2:
                  description: Пример неправильного запроса
                  value:
                    message: Отрицательное значение в кол-ве осадков
                    code: 400
        '401':
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: Пример ответа
                  value:
                    message: Пользователь не авторизован
                    code: 401
        '403':
          description: Пользователю не хватает прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: Пример ответа
                  value:
                    message: Недостаточно прав для выполнения операции
                    code: 403
        '500':
          description: Ошибка на сервере
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                example1:
                  description: Пример ответа на 500
                  value:
                    message: Произошла ошибка на стороне сервера
                    code: 500
    patch:
      tags:
        - PrecipitationRecordController
      summary: Обновить запись по осадкам
      description: Измененение минимальной, максимальной границы и кол-во осадков
      operationId: patchRecord
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrecipitationRecordUpdateDto'
      responses:
        '200':
          description: Успешное обновление записи
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrecipitationRecordResponseDto'
        '400':
          description: Ошибка в составлении запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                example1:
                  description: Пример неправильного запроса
                  value:
                    message: Минимальная граница больше максимальной
                    code: 400
                example2:
                  description: Пример неправильного запроса
                  value:
                    message: Отрицательное значение в кол-ве осадков
                    code: 400
        '401':
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: Пример ответа
                  value:
                    message: Пользователь не авторизован
                    code: 401
        '403':
          description: Пользователю не хватает прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                exmaple1:
                  description: Пример ответа
                  value:
                    message: Недостаточно прав для выполнения операции
                    code: 403
        '500':
          description: Ошибка на сервере
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiExceptionDto'
              examples:
                example1:
                  description: Пример ответа на 500
                  value:
                    message: Произошла ошибка на стороне сервера
                    code: 500
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiExceptionDto:
      type: object
      description: ДТО для ответа с ошибкой
      required:
        - code
        - message
      properties:
        message:
          type: string
          description: Сообщение с ошибкой
        code:
          type: integer
          format: int32
    # DTO для погодных параметров
...
    FFWIResponseDto:
      description: DTO представления КППО в приложении
      properties:
        id:
          type: string
          description: ID КППО
          format: uuid
          example: 92eb322f-9fec-4f5c-adbf-008627ed271a
        name:
          type: string
          description: Название КППО
          example: КППО Нестерова
        formula:
          type: string
          description: Формула определения КП
          example: ($T_d - $T_0) * $T_d
        fwis:
          type: string
          description: Список названий КПО, придлежащих КППО
          example: КПО_Нестеров_1, КПО_Нестеров_2, КПО_Нестеров_3
        weatherDataParams:
          type: array
          items:
            $ref: '#/components/schemas/FFWIWeatherParamDto'
        precipitationRecords:
          type: array
          items:
            $ref: '#/components/schemas/PrecipitationRecordResponseDto'
    FFWIWeatherParamDto:
      description: Представление погодного параметра с параметром, связывающим формулу
      properties:
        weatherDataId:
          type: string
          description: ID погодного параметра
          example: 2a3c6c0e-2d26-49fb-ab45-55a3e703cace
        weatherName:
          type: string
          description: Название погодного условия
          example: Температура воздуха
        formulaParameter:
          type: string
          description: Название параметра в формуле
          example: $T_d
    FFWICreateRequestDto:
      description: DTO для создания КППО
      required:
        - name
        - formula
        - weatherDataIds
      properties:
        name:
          type: string
          description: Название КППО
          example: КППО Свердловой
        formula:
          type: string
          description: Формула определения КП
        weatherDataParams:
          type: array
          items:
            $ref: '#/components/schemas/FFWIWeatherParamDto'
    # DTO для КПО
    FWIResponseDto:
      description: DTO представление КПО в приложении
      properties:
        id:
          type: string
          format: uuid
          example: 12efb6fa-72b7-455e-87c2-872d24d1f22d
        name:
          type: string
          example: КПО_Нестерова_1
        minValue:
          type: number
          description: нижняя граница горммости
          minimum: 0
          example: 0
        maxValue:
          type: number
          description: верхняя граница горимости
          minimum: 0
          example: 1000
        fireDangerName:
          type: string
          description: Имя уровня пожарной угрозы
          example: Незначительный
        ffwiName:
          type: string
          description: Название КППО
          example: КППО_Нестерова
...
    PrecipitationRecordResponseDto:
      properties:
        id:
          type: integer
          format: int64
          example: 1
        minValue:
          type: number
          description: Минимальная граница горимости
          example: 0
        maxValue:
          type: number
          description: Максимальная граница горимости
          example: 100000.0
        value:
          type: integer
          description: Значение осадков в мм
          example: 3
        ffwiName:
          type: string
          description: Название КППО
          example: КППО_Нестерова
        ffwId:
          type: string
          format: uuid
          description: ID КППО
          example: 97b4e98f-6829-446a-aea7-8e7be17d56c9
    PrecipitationRecordCreateDto:
      required:
        - minValue
        - maxValue
        - value
        - ffwi_id
      properties:
        minValue:
          type: number
          description: Минимальная граница горимости
          example: 0
        maxValue:
          type: number
          description: Максимальная граница горимости
          example: 100000.0
        value:
          type: integer
          description: Значение осадков в мм
          example: 3
        ffwi_id:
          type: string
          format: uuid
          description: ID КППО
          example: 97b4e98f-6829-446a-aea7-8e7be17d56c9
    PrecipitationRecordUpdateDto:
      required:
        - id
      properties:
        id:
          type: integer
          format: int64
          example: 1
        minValue:
          type: number
          description: Минимальная граница горимости
          example: 0
        maxValue:
          type: number
          description: Максимальная граница горимости
          example: 100000.0
        value:
          type: integer
          description: Значение осадков в мм
          example: 3

    # DTO для пользователя
...